Add tests and rename files
