+++ src/taktyk/main.py
@@ -8,8 +8,8 @@ from typing import NoReturn, Tuple, List
 
 from wykop import WykopAPI, MultiKeyWykopAPI
 
-from taktyk.ReminderRepository import ShelveReminderRepository
-from taktyk.TaktykBot import TaktykBot
+from taktyk.reminder_repository import ShelveReminderRepository
+from taktyk.taktyk_bot import TaktykBot
 
 KEYS_FILE_NAME = 'keys'
 WYKOP_APP_KEY = 'aNd401dAPp'
