+++ src/taktyk/taktyk_bot.py
@@ -3,7 +3,7 @@ from wykop import WykopAPI, WykopAPIError
 
 import logging
 
-from taktyk.ReminderRepository import ReminderRepository
+from taktyk.reminder_repository import ReminderRepository
 from taktyk.model import Reminder, entry_url
