+++ src/tests/test_taktyk.py
@@ -1,9 +1,8 @@
-from taktyk.ReminderRepository import InMemoryReminderRepository
-from taktyk.TaktykBot import TaktykBot
+from taktyk.reminder_repository import InMemoryReminderRepository
+from taktyk.taktyk_bot import TaktykBot
 from tests.FakeWykopApi import FakeWykopApi
 
 
-# Fix impost
 class TestTaktyk(object):
 
     def test_get_reminders_from_more_then_one_page(self):
@@ -73,3 +72,39 @@ class TestTaktyk(object):
         bot.send_reminders()
 
         assert len(api.get_sent_messages()[login]) == 1
+
+    def test_send_message_with_link_to_last_read_comment(self):
+        entry_id = 'id-1'
+        start_comments_count = 1
+        login = 'login1'
+        repository = InMemoryReminderRepository()
+        api = FakeWykopApi()
+        bot = TaktykBot(api, repository)
+
+        api.add_entry(entry_id, start_comments_count)
+        api.add_notification(login, entry_id, 'sub-id-1', 1)
+        bot.save_new_reminders()
+
+        api.set_entry_comments_count(entry_id, start_comments_count + 2)
+        bot.send_reminders()
+
+        assert 'https://www.wykop.pl/wpis/id-1/#comment-sub-id-1' in api.get_sent_messages()[login][0]
+
+    def test_send_second_message_with_link_to_last_read_comment(self):
+        entry_id = 'id-1'
+        start_comments_count = 1
+        login = 'login1'
+        repository = InMemoryReminderRepository()
+        api = FakeWykopApi()
+        bot = TaktykBot(api, repository)
+
+        api.add_entry(entry_id, start_comments_count)
+        api.add_notification(login, entry_id, 'sub-id-1', 1)
+        bot.save_new_reminders()
+        api.set_entry_comments_count(entry_id, start_comments_count + 1)
+        bot.send_reminders()
+
+        api.set_entry_comments_count(entry_id, start_comments_count + 2)
+        bot.send_reminders()
+
+        assert 'https://www.wykop.pl/wpis/id-1/#comment-sub-id-3' in api.get_sent_messages()[login][-1]
